+++ experiments/cluster/ardqn_post_experiment.py
@@ -1,10 +1,10 @@
 import argparse
-import warnings
 from os import path
 
 from experiments.custom_envs import ENV_DICT
 from sb3_contrib import ARDQN
 from sb3_contrib.common.satisficing.evaluation import plot_ar
+import pickle
 
 
 def load_model(log_path, name):
@@ -13,21 +13,30 @@ def load_model(log_path, name):
     return model
 
 
+def load_results(log_path, name):
+    with open(path.join(log_path, "ARDQN", name, "results.pkl"), "rb") as f:
+        results = pickle.load(f)
+    return results
+
+
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument("--names", nargs="+", type=str, required=True)
-    parser.add_argument("--env_id", type=str, required=True)
     parser.add_argument("--log_path", type=str, required=True)
-    parser.add_argument("--n_eval_episodes", type=int, default=100)
     args = parser.parse_args()
-    env = ENV_DICT[args.env_id]()
-    models = []
-    for name in args.names:
-        if path.exists(path.join(args.log_path, "ARDQN", name, "models")):
-            models.append(load_model(args.log_path, name))
-        else:
-            warnings.warn(f"Model {name} not found in {args.log_path}")
-    print(f"Loaded {len(models)} models")
-    fig = plot_ar(models, env=env, n_eval_episodes=args.n_eval_episodes)
+    models = list(
+        map(
+            load_model,
+            args.names,
+        )
+    )
+    results = list(
+        map(
+            load_results,
+            args.names,
+        )
+    )
+
+    fig = plot_ar(models, results=results)
     # Save plotly figure as html
     fig.write_html(path.join(args.log_path, "ARDQN", "results.html"))
