+++ tests/test_save_load.py
@@ -29,8 +29,8 @@ def select_env(model_class: BaseAlgorithm) -> gym.Env:
 
 
 @pytest.mark.parametrize("model_class", MODEL_LIST)
-@pytest.mark.parametrize("shared_network", ["all", "none", "features_extractor"])
-def test_save_load(tmp_path, model_class, shared_network):
+@pytest.mark.parametrize("shared_network", ["all", "none", "features_extractor", "min_max"])
+def test_save_load(tmp_path, model_class, shared_network):  # todo: make this pass for "all" and "min_max"
     """
     Test if 'save' and 'load' saves and loads model correctly
     and if 'get_parameters' and 'set_parameters' and work correctly.
@@ -39,7 +39,8 @@ def test_save_load(tmp_path, model_class, shared_network):
 
     :param model_class: (BaseAlgorithm) A RL model
     """
-
+    if shared_network != "all" and model_class != ARDQN:
+        pytest.skip()
     env = DummyVecEnv([lambda: select_env(model_class)])
 
     policy_kwargs = dict(net_arch=[16])
@@ -206,7 +207,7 @@ def test_set_env(model_class):
 
 
 @pytest.mark.parametrize("model_class", MODEL_LIST)
-@pytest.mark.parametrize("shared_network", ["all", "none", "features_extractor"])
+@pytest.mark.parametrize("shared_network", ["all", "features_extractor", "none", "min_max"])
 def test_exclude_include_saved_params(tmp_path, model_class, shared_network):
     """
     Test if exclude and include parameters of save() work
@@ -270,7 +271,7 @@ def test_save_load_replay_buffer(tmp_path, model_class):
 
 @pytest.mark.parametrize("model_class", MODEL_LIST)
 @pytest.mark.parametrize("policy_str", ["MlpPolicy", "CnnPolicy"])
-@pytest.mark.parametrize("shared_network", ["all", "none", "features_extractor"])
+@pytest.mark.parametrize("shared_network", ["all", "features_extractor", "none", "min_max"])
 def test_save_load_policy(tmp_path, model_class, policy_str, shared_network):
     """
     Test saving and loading policy only.
